requirements-dev.txt
factory-boy==3.3.0
pip-tools
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
